edge/requirements.txt
sense-hat
paho-mqtt
pymodbus>=3.4
Flask